pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "corpus"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "sqlalchemy>=2.0",
    "sentence-transformers",
    "datasketch",
    "numpy",
    "chromadb",
    "networkx",
    "click",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "openai",
    "watchdog>=3.0",
    "beautifulsoup4>=4.12",
    "pymupdf>=1.24",
]

[project.optional-dependencies]
git = ["gitpython>=3.1"]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "pytest-asyncio",
    "ruff",
    "mypy",
]

[project.scripts]
corpus = "corpus.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
markers = [
    "integration: marks tests that call external services (deselect with '-m \"not integration\"')",
]

[tool.ruff]
target-version = "py312"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W"]

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
ignore_missing_imports = true
